requirements.txt
aeneas==1.7.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
click==8.1.8
Flask==3.1.1
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==5.4.0
MarkupSafe==3.0.2
numpy==2.0.2
praat-parselmouth==0.4.6
soupsieve==2.7
typing_extensions==4.14.0
Werkzeug==3.1.3
zipp==3.23.0
pypinyin
opencc-python-reimplemented
aeneas
matplotlib